+++ src/globals.cpp
@@ -6,7 +6,7 @@ Point::Point()
     y = 0;
 }
 
-Point::Point(int _x, int _y)
+Point::Point(float _x, float _y)
 {
     x = _x;
     y = _y;
